implement export as rdf
First pass at exporting as rdf. The button is currently the new download
button at the top which I enabled. Currently I'm using a kind of clunky
ajax request to the page because it allows me to control the filename.
Simply using window.open() on the force-open link is much faster but I
couldn't find a way to control filename with that.